+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Apimap - Send engine name, engine version, framework name, framework version, database version and orm name in apimap meta.
 
 ## RELEASE 3.0.0-beta.3 - 2019-02-08
 ### Changed